fix(sdk): allow scrolling in embed during payment flow
