+++ packages/sdk/src/react/boosts/BoostLoader.tsx
@@ -1,29 +1,17 @@
 import * as React from "react";
-import styled, { createGlobalStyle, ThemeProvider } from "styled-components";
+import styled, { ThemeProvider } from "styled-components";
 import { withRouter, RouteComponentProps } from "react-router";
 import {
   CivilContext,
   ICivilContext,
   RENDER_CONTEXT,
   CivilIcon,
-  ChevronAnchor,
   DEFAULT_BUTTON_THEME,
   DEFAULT_CHECKBOX_THEME,
 } from "@joincivil/components";
-import { colors, mediaQueries } from "@joincivil/elements";
+import { mediaQueries } from "@joincivil/elements";
 import { Boost } from "./Boost";
 
-const GlobalStyleNoScroll = createGlobalStyle`
-  html, body {
-    height: 100%;
-    overflow: hidden;
-  }
-  body {
-    border: 1px solid ${colors.accent.CIVIL_GRAY_4};
-    margin: 1px; // otherwise border can be clipped by inner edge of iframe
-    border-radius: 4px;
-  }
-`;
 const CivilLogoLink = styled.a`
   position: absolute;
   display: inline-block;
@@ -47,35 +35,6 @@ const CivilLogoLink = styled.a`
     height: auto;
   }
 `;
-const OverflowLinkContainer = styled.div`
-  position: absolute;
-  z-index: 1;
-  width: calc(100% - 2px);
-  height: 100px;
-  bottom: 1px;
-  left: 1px;
-  background: rgb(255, 255, 255);
-  background: linear-gradient(
-    180deg,
-    rgba(255, 255, 255, 0) 0%,
-    rgba(255, 255, 255, 0.85) 25%,
-    rgba(255, 255, 255, 1) 100%
-  );
-  border-radius: 4px;
-`;
-const OverflowLink = styled(ChevronAnchor)`
-  position: absolute;
-  display: block;
-  width: 100%;
-  bottom: 0;
-  padding: 16px 0;
-  background: white;
-  border-top: 1px solid ${colors.accent.CIVIL_GRAY_4};
-  text-align: center;
-  &:hover {
-    text-decoration: underline;
-  }
-`;
 
 export interface BoostLoaderParams {
   boostId: string;
@@ -94,15 +53,9 @@ const BoostLoaderComponent = (props: RouteComponentProps<BoostLoaderParams>) => 
 
   return (
     <>
-      <GlobalStyleNoScroll />
       <CivilLogoLink href="https://registry.civil.co" target="_blank">
         <CivilIcon />
       </CivilLogoLink>
-      <OverflowLinkContainer>
-        <OverflowLink href={"https://registry.civil.co/boosts/" + props.match.params.boostId} target="_blank">
-          View More
-        </OverflowLink>
-      </OverflowLinkContainer>
       <ThemeProvider theme={theme}>
         <Boost boostId={props.match.params.boostId} open={true} payment={!!props.match.params.payment} />
       </ThemeProvider>
